move all directives to seperate files
