+++ npm/ng-packs/packages/theme-shared/src/lib/components/card/card-header.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from "@angular/core";
+
+@Directive({
+  selector: `abp-card-header, [abp-card-header], [abpCardHeader]`,
+  host: {
+    class: 'card-header',
+  },
+})
+export class CardHeader {}
