+++ npm/ng-packs/packages/theme-shared/src/lib/components/card/card-img-top.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from "@angular/core";
+
+@Directive({
+  selector: `abp-card-img-top, [abp-card-img-top], [abpCardImgTop]`,
+  host: {
+    class: 'card-img-top',
+  },
+})
+export class CardImgTop {}
