+++ npm/ng-packs/packages/theme-shared/src/lib/components/card/card-subtitle.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from "@angular/core";
+
+@Directive({
+  selector: `abp-card-subtitle, [abp-card-subtitle], [abpCardSubtitle]`,
+  host: {
+    class: 'card-subtitle',
+  },
+})
+export class CardSubtitle {}
