+++ npm/ng-packs/packages/theme-shared/src/lib/components/card/card-title.directive.ts
@@ -1,0 +1,9 @@
+import { Directive } from "@angular/core";
+
+@Directive({
+  selector: `abp-card-title, [abp-card-title], [abpCardTitle]`,
+  host: {
+    class: 'card-title',
+  },
+})
+export class CardTitle {}
