+++ npm/ng-packs/packages/theme-shared/src/lib/components/card/card.module.ts
@@ -1,17 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { CardBodyComponent } from './card-body.component';
-import {
-  CardComponent,
-  CardImgTop,
-  CardLink,
-  CardSubtitle,
-  CardText,
-  CardTitle,
-  CardHeader,
-} from './card.component';
+import { CardComponent } from './card.component';
 import { CardHeaderComponent } from './card-header.component';
 import { CardFooterComponent } from './card-footer.component';
+import { CardTitle } from './card-title.directive';
+import { CardSubtitle } from './card-subtitle.directive';
+import { CardImgTop } from './card-img-top.directive';
+import { CardHeader } from './card-header.directive';
 
 const declarationsWithExports = [
   CardComponent,
@@ -20,9 +16,7 @@ const declarationsWithExports = [
   CardFooterComponent,
   CardTitle,
   CardSubtitle,
-  CardText,
   CardImgTop,
-  CardLink,
   CardHeader,
 ];
 
@@ -31,4 +25,4 @@ const declarationsWithExports = [
   imports: [CommonModule],
   exports: [...declarationsWithExports],
 })
-export class CardModule {}
+export class CardModule { }
