+++ npm/ng-packs/packages/theme-shared/src/lib/components/card/index.ts
@@ -3,3 +3,7 @@ export * from './card.component';
 export * from './card-body.component';
 export * from './card-footer.component';
 export * from './card-header.component';
+export * from './card-title.directive';
+export * from './card-subtitle.directive';
+export * from './card-img-top.directive';
+export * from './card-header.directive';
